requirements.txt
geoclip
torch
torchvision
datasets
Pillow
pandas
numpy
pyyaml
tqdm
scikit-learn
matplotlib
